Update from mcp_server to wandb_mcp_server
