+++ src/wandb_mcp_server/query_wandbot.py
@@ -1,0 +1,7 @@
+"""
+Import the query_wandbot_api function from the wandbot package, a helper function
+to query the Weights & Biases support bot.
+"""
+
+from wandbot.wandbot_tool import query_wandbot_api, WANDBOT_TOOL_DESCRIPTION
+
